# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SafarSathi carpooling API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterInputGender = typeof RegisterInputGender[keyof typeof RegisterInputGender];


export const RegisterInputGender = {
  M: 'M',
  F: 'F',
  O: 'O',
} as const;

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  email: string;
  phone_number: string;
  university: string;
  gender: RegisterInputGender;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UserProfileGender = typeof UserProfileGender[keyof typeof UserProfileGender];


export const UserProfileGender = {
  M: 'M',
  F: 'F',
  O: 'O',
} as const;

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  phone_number: string;
  university: string;
  gender: UserProfileGender;
  /** @nullable */
  avatar_url?: string | null;
  created_at?: string;
}

export interface AuthResponse {
  user: UserProfile;
  token: string;
}

export interface UpdateProfileInput {
  /** @nullable */
  avatar_url?: string | null;
  phone_number?: string;
  email?: string;
  university?: string;
}

export type RideGenderPreference = typeof RideGenderPreference[keyof typeof RideGenderPreference];


export const RideGenderPreference = {
  ANY: 'ANY',
  MALE: 'MALE',
  FEMALE: 'FEMALE',
} as const;

export type RideStatus = typeof RideStatus[keyof typeof RideStatus];


export const RideStatus = {
  OPEN: 'OPEN',
  FULL: 'FULL',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
} as const;

export interface Ride {
  id: number;
  driver_id: number;
  driver_name: string;
  driver_university?: string;
  /** @nullable */
  driver_avg_rating?: number | null;
  origin: string;
  destination: string;
  /** @nullable */
  origin_lat?: number | null;
  /** @nullable */
  origin_lng?: number | null;
  /** @nullable */
  dest_lat?: number | null;
  /** @nullable */
  dest_lng?: number | null;
  departure_time: string;
  available_seats: number;
  fare: number;
  transport_type: string;
  gender_preference: RideGenderPreference;
  status: RideStatus;
  request_count?: number;
  created_at: string;
}

export type RideInputGenderPreference = typeof RideInputGenderPreference[keyof typeof RideInputGenderPreference];


export const RideInputGenderPreference = {
  ANY: 'ANY',
  MALE: 'MALE',
  FEMALE: 'FEMALE',
} as const;

export interface RideInput {
  /** @minLength 1 */
  origin: string;
  /** @minLength 1 */
  destination: string;
  /** @nullable */
  origin_lat?: number | null;
  /** @nullable */
  origin_lng?: number | null;
  /** @nullable */
  dest_lat?: number | null;
  /** @nullable */
  dest_lng?: number | null;
  departure_time: string;
  /** @minimum 1 */
  available_seats: number;
  /** @minimum 0 */
  fare: number;
  transport_type?: string;
  gender_preference?: RideInputGenderPreference;
}

export type RideUpdateGenderPreference = typeof RideUpdateGenderPreference[keyof typeof RideUpdateGenderPreference];


export const RideUpdateGenderPreference = {
  ANY: 'ANY',
  MALE: 'MALE',
  FEMALE: 'FEMALE',
} as const;

export type RideUpdateStatus = typeof RideUpdateStatus[keyof typeof RideUpdateStatus];


export const RideUpdateStatus = {
  OPEN: 'OPEN',
  FULL: 'FULL',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
} as const;

export interface RideUpdate {
  origin?: string;
  destination?: string;
  /** @nullable */
  origin_lat?: number | null;
  /** @nullable */
  origin_lng?: number | null;
  /** @nullable */
  dest_lat?: number | null;
  /** @nullable */
  dest_lng?: number | null;
  departure_time?: string;
  available_seats?: number;
  fare?: number;
  transport_type?: string;
  gender_preference?: RideUpdateGenderPreference;
  status?: RideUpdateStatus;
}

export type RideRequestStatus = typeof RideRequestStatus[keyof typeof RideRequestStatus];


export const RideRequestStatus = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED',
} as const;

export interface RideRequest {
  id: number;
  ride_id: number;
  rider_id: number;
  rider_name: string;
  rider_university?: string;
  rider_gender?: string;
  status: RideRequestStatus;
  /**
     * Revealed to the rider only when request is ACCEPTED
     * @nullable
     */
  driver_phone?: string | null;
  /**
     * Revealed to the driver only when request is ACCEPTED
     * @nullable
     */
  rider_phone?: string | null;
  ride?: Ride;
  created_at: string;
}

export interface RideRequestInput {
  ride_id: number;
}

export type RideRequestUpdateStatus = typeof RideRequestUpdateStatus[keyof typeof RideRequestUpdateStatus];


export const RideRequestUpdateStatus = {
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED',
} as const;

export interface RideRequestUpdate {
  status: RideRequestUpdateStatus;
}

export interface Review {
  id: number;
  reviewer_id: number;
  reviewer_name: string;
  driver_id: number;
  ride_id: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comment?: string | null;
  created_at: string;
}

export interface ReviewInput {
  driver_id: number;
  ride_id: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment?: string;
}

export interface UserReviews {
  user_id: number;
  /** @nullable */
  avg_rating: number | null;
  review_count: number;
  reviews: Review[];
}

export interface PopularRoute {
  origin: string;
  destination: string;
  count: number;
}

export interface StatsSummary {
  total_open_rides: number;
  total_seats_available: number;
  my_rides_count: number;
  my_requests_count: number;
  my_pending_requests: number;
  popular_routes?: PopularRoute[];
}

export type DeleteAccount200 = {
  ok?: boolean;
};

export type ListRidesParams = {
origin?: string;
destination?: string;
gender_preference?: ListRidesGenderPreference;
/**
 * Filter by departure date (YYYY-MM-DD)
 */
date?: string;
/**
 * Search origin latitude for proximity sort
 */
origin_lat?: number;
/**
 * Search origin longitude for proximity sort
 */
origin_lng?: number;
/**
 * Search destination latitude for proximity sort
 */
dest_lat?: number;
/**
 * Search destination longitude for proximity sort
 */
dest_lng?: number;
};

export type ListRidesGenderPreference = typeof ListRidesGenderPreference[keyof typeof ListRidesGenderPreference];


export const ListRidesGenderPreference = {
  ANY: 'ANY',
  MALE: 'MALE',
  FEMALE: 'FEMALE',
} as const;