# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SafarSathi carpooling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user profile (avatar etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    delete:
      operationId: deleteAccount
      tags: [auth]
      summary: Close/delete current user account
      responses:
        "200":
          description: Account closed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [username, password, email, phone_number, university, gender]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        email:
          type: string
        phone_number:
          type: string
        university:
          type: string
        gender:
          type: string
          enum: [M, F, O]
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string
    UserProfile:
      type: object
      required: [id, username, email, phone_number, university, gender]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        phone_number:
          type: string
        university:
          type: string
        gender:
          type: string
          enum: [M, F, O]
        avatar_url:
          type: ["string", "null"]
        created_at:
          type: string
    UpdateProfileInput:
      type: object
      properties:
        avatar_url:
          type: ["string", "null"]
        phone_number:
          type: string
        email:
          type: string
        university:
          type: string
...