# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SafarSathi carpooling API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterInputGender } from './registerInputGender';

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  email: string;
  phone_number: string;
  university: string;
  gender: RegisterInputGender;
}