# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SafarSathi carpooling API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileGender } from './userProfileGender';

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  phone_number: string;
  university: string;
  gender: UserProfileGender;
  /** @nullable */
  avatar_url?: string | null;
  created_at?: string;
}